eda-venus-change-detection/requirements.txt
lightning>=2.0
jsonargparse[signatures]
jupyterlab
matplotlib
rasterio
albumentations
torchvision
torchmetrics>=1.2.0
wandb